# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Dojiva API - gamified trading learning
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /simulator/trades:
    post:
      operationId: submitTrade
      tags: [trading]
      summary: Submit a trade decision and advance the market
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeInput"
      responses:
        "200":
          description: Trade result with revealed candles and feedback
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeResult"
        "404":
          description: Scenario not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
...
  /strategies:
    get:
      operationId: listStrategies
      tags: [trading]
      summary: List saved strategies
      responses:
        "200":
          description: Strategies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Strategy"
    post:
      operationId: createStrategy
      tags: [trading]
      summary: Create a strategy from the builder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StrategyInput"
      responses:
        "200":
          description: Created strategy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Strategy"
  /strategies/{id}:
    delete:
      operationId: deleteStrategy
      tags: [trading]
      summary: Delete a strategy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkMessage"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
...
components:
  schemas:
...
    ApiMessage:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    OkMessage:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
...
    Candle:
      type: object
      required: [o, h, l, c]
      properties:
        o:
          type: number
        h:
          type: number
        l:
          type: number
        c:
          type: number
...
    TradeInput:
      type: object
      required: [scenarioId, direction]
      properties:
        scenarioId:
          type: integer
        direction:
          type: string
          enum: [buy, sell, wait]
        entry:
          type: ["number", "null"]
        stopLoss:
          type: ["number", "null"]
        takeProfit:
          type: ["number", "null"]
        riskPercent:
          type: ["number", "null"]
          description: Percent of balance risked (default 1)
        emotion:
          type: ["string", "null"]
          description: How the user felt (confiant, stressé, neutre...)
        strategyId:
          type: ["integer", "null"]
    TradeResult:
      type: object
      required: [outcome, pnl, balance, futureCandles, feedback, riskReward]
      properties:
        outcome:
          type: string
          enum: [take-profit, stop-loss, expired, waited]
        pnl:
          type: number
        balance:
          type: number
        exitPrice:
          type: ["number", "null"]
        riskReward:
          type: ["number", "null"]
        futureCandles:
          type: array
          items:
            $ref: "#/components/schemas/Candle"
        feedback:
          type: array
          description: Pedagogical feedback lines in French
          items:
            type: string
...
    StrategyInput:
      type: object
      required: [name, market, style, timeframe, context, entryRules, riskPercent]
      properties:
        name:
          type: string
        market:
          type: string
        style:
          type: string
        timeframe:
          type: string
        context:
          type: array
          items:
            type: string
        entryRules:
          type: array
          items:
            type: string
        stopLossRule:
          type: ["string", "null"]
        takeProfitRule:
          type: ["string", "null"]
        riskPercent:
          type: number
    Strategy:
      type: object
      required: [id, name, market, style, timeframe, context, entryRules, riskPercent, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        market:
          type: string
        style:
          type: string
        timeframe:
          type: string
        context:
          type: array
          items:
            type: string
        entryRules:
          type: array
          items:
            type: string
        stopLossRule:
          type: ["string", "null"]
        takeProfitRule:
          type: ["string", "null"]
        riskPercent:
          type: number
        createdAt:
          type: string
...