# lib/api-zod/src/generated/types/candle.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Dojiva API - gamified trading learning
 * OpenAPI spec version: 0.1.0
 */

export interface Candle {
  o: number;
  h: number;
  l: number;
  c: number;
}