# lib/api-zod/src/generated/types/strategy.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Dojiva API - gamified trading learning
 * OpenAPI spec version: 0.1.0
 */

export interface Strategy {
  id: number;
  name: string;
  market: string;
  style: string;
  timeframe: string;
  context: string[];
  entryRules: string[];
  /** @nullable */
  stopLossRule?: string | null;
  /** @nullable */
  takeProfitRule?: string | null;
  riskPercent: number;
  createdAt: string;
}