# lib/api-zod/src/generated/types/tradeInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Dojiva API - gamified trading learning
 * OpenAPI spec version: 0.1.0
 */
import type { TradeInputDirection } from './tradeInputDirection';

export interface TradeInput {
  scenarioId: number;
  direction: TradeInputDirection;
  /** @nullable */
  entry?: number | null;
  /** @nullable */
  stopLoss?: number | null;
  /** @nullable */
  takeProfit?: number | null;
  /**
     * Percent of balance risked (default 1)
     * @nullable
     */
  riskPercent?: number | null;
  /**
     * How the user felt (confiant, stressé, neutre...)
     * @nullable
     */
  emotion?: string | null;
  /** @nullable */
  strategyId?: number | null;
}