# lib/api-zod/src/generated/types/tradeResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Dojiva API - gamified trading learning
 * OpenAPI spec version: 0.1.0
 */
import type { Candle } from './candle';
import type { TradeResultOutcome } from './tradeResultOutcome';

export interface TradeResult {
  outcome: TradeResultOutcome;
  pnl: number;
  balance: number;
  /** @nullable */
  exitPrice?: number | null;
  /** @nullable */
  riskReward: number | null;
  futureCandles: Candle[];
  /** Pedagogical feedback lines in French */
  feedback: string[];
}